refactoring robot_arm -> robot_arm_client
