+++ robot_arm/robot_arm_client/setup.cfg
@@ -1,4 +1,4 @@
 [develop]
-script_dir=$base/lib/robot_arm
+script_dir=$base/lib/robot_arm_client
 [install]
-install_scripts=$base/lib/robot_arm
+install_scripts=$base/lib/robot_arm_client
