+++ robot_arm/robot_arm_client/setup.py
@@ -1,25 +1,23 @@
 from setuptools import find_packages, setup
 
-package_name = 'robot_arm'
+package_name = "robot_arm_client"
 
 setup(
     name=package_name,
-    version='0.0.0',
-    packages=find_packages(exclude=['test']),
+    version="0.0.0",
+    packages=find_packages(exclude=["test"]),
     data_files=[
-        ('share/ament_index/resource_index/packages',
-            ['resource/' + package_name]),
-        ('share/' + package_name, ['package.xml']),
+        ("share/ament_index/resource_index/packages", ["resource/" + package_name]),
+        ("share/" + package_name, ["package.xml"]),
     ],
-    install_requires=['setuptools'],
+    install_requires=["setuptools"],
     zip_safe=True,
-    maintainer='user',
-    maintainer_email='charles.sirois@mail.mcgill.ca',
-    description='TODO: Package description',
-    license='Apache-2.0',
-    tests_require=['pytest'],
+    maintainer="user",
+    maintainer_email="charles.sirois@mail.mcgill.ca",
+    description="TODO: Package description",
+    license="Apache-2.0",
+    tests_require=["pytest"],
     entry_points={
-        'console_scripts': [
-        ],
+        "console_scripts": [],
     },
 )
